src/backend/data/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Flask==3.0.0
Flask-Cors==4.0.0
google-api-core==2.15.0
google-auth==2.26.1
google-cloud-appengine-logging==1.4.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
parso==0.8.3
pluggy==1.3.0
proto-plus==1.23.0
protobuf==4.25.2
pudb==2023.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pytest==7.4.4
pytest-pudb==0.7.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
urllib3==2.1.0
urwid==2.4.1
urwid-readline==0.13
Werkzeug==3.0.1
